q3_strategic_decision_engine/requirements.txt
# Core Framework Dependencies
fastapi>=0.104.0,<0.115.0
uvicorn[standard]>=0.24.0,<0.25.0
pydantic>=2.5.0,<3.0.0
streamlit>=1.28.0,<2.0.0

# LangChain and AI Dependencies
langchain>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0
langchain-google-genai>=2.0.0,<3.0.0
langchain-anthropic>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-experimental>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0

# Vector Database and Embeddings
chromadb>=0.4.15,<0.5.0
sentence-transformers>=2.2.0,<3.0.0
faiss-cpu>=1.7.0,<2.0.0

# Database Dependencies
psycopg2-binary>=2.9.0,<3.0.0
sqlalchemy>=2.0.0,<3.0.0
redis>=5.0.0,<6.0.0
pymongo>=4.6.0,<5.0.0

# Document Processing
pypdf>=3.17.0,<4.0.0
python-docx>=1.1.0,<2.0.0
python-pptx>=0.6.0,<1.0.0
openpyxl>=3.1.0,<4.0.0
pandas>=2.1.0,<3.0.0

# RAG and Evaluation
ragas>=0.1.0,<0.2.0
rank-bm25>=0.2.0,<0.3.0

# Visualization and Charts
plotly>=5.17.0,<6.0.0
matplotlib>=3.8.0,<4.0.0
seaborn>=0.13.0,<0.14.0
altair>=5.2.0,<6.0.0

# API and External Services
requests>=2.31.0,<3.0.0
httpx>=0.25.0,<0.26.0
yfinance>=0.2.0,<0.3.0
alpha-vantage>=2.3.0,<3.0.0

# Utilities and Helpers
python-multipart>=0.0.6,<0.1.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic-settings>=2.1.0,<3.0.0

# Async and Concurrency
asyncio-mqtt>=0.16.0,<0.17.0
aiofiles>=23.2.0,<24.0.0
aioredis>=2.0.0,<3.0.0

# Development and Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<0.22.0
pytest-cov>=4.1.0,<5.0.0
pytest-mock>=3.12.0,<4.0.0
factory-boy>=3.3.0,<4.0.0
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.7.0,<2.0.0

# Security
bcrypt>=4.1.0,<5.0.0
cryptography>=41.0.0,<42.0.0

# Additional AI and ML
torch>=2.1.0,<3.0.0
transformers>=4.35.0,<5.0.0
scikit-learn>=1.3.0,<2.0.0
numpy>=1.24.0,<2.0.0
scipy>=1.11.0,<2.0.0

# Financial and Market Data
fredapi>=0.5.0,<0.6.0
quandl>=3.7.0,<4.0.0

# Monitoring and Logging
structlog>=23.2.0,<24.0.0
prometheus-client>=0.19.0,<0.20.0
psutil>=5.9.0,<6.0.0 